# service/api-java/docs/openapi.yml
openapi: 3.0.3
info:
  title: GreenTech API
  version: "1.0.0"
  description: >
    Authentication API for the community waste-management platform. Local email/password
    signup goes through a two-step OTP flow; Google OAuth is also supported.


    ## Tokens

    Authentication returns a single `accessToken` (JWT, 7 days) sent back as
    `Authorization: Bearer <token>`. There are no refresh tokens — when the access token
    expires the client logs in again. The token carries the user id as `sub` and the role
    as a `role` claim, but clients should read role and points from the `user` object
    rather than decoding the JWT.


    ## Roles

    Every account has exactly one role, fixed at signup. `CITIZEN`, `COLLECTOR` and
    `RECYCLER` can be chosen by the caller; `MUNICIPAL_ADMIN` and `SUPER_ADMIN` are
    rejected with 403 and must be assigned directly in the database.


    ## Errors

    Every failure returns `{"error": "<message>"}` with the relevant HTTP status. Validation
    failures join all field errors into one string, e.g. `"password: size must be between
    8 and 72"`.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/listings:
    post:
      tags: [Marketplace]
      summary: List waste for sale
      description: >
        Attach a `detectionId` and the material, weight, image and description are copied from
        that scan. Without one, supply `material` and `weightKg` yourself. A scan can only have
        one open listing at a time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          description: Missing material/weight, non-positive price, or an ineligible scan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Scan not found or not yours
        '409':
          description: That scan is already listed

    get:
      tags: [Marketplace]
      summary: Browse open listings
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Open listings, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/listings/mine:
    get:
      tags: [Marketplace]
      summary: My listings or my purchases
      description: >
        Scoped by role. A recycler sees what they bought; everyone else sees what they listed.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Listings for this caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/listings/{id}:
    get:
      tags: [Marketplace]
      summary: One listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Listing not found
  /api/v1/listings/{id}/interested:
    post:
      tags: [Marketplace]
      summary: Buy a listing (recycler)
      description: >
        Recyclers only. Claims the listing atomically, marks it SOLD, credits the seller and
        debits the buyer. If several recyclers tap at the same instant exactly one gets 200 and
        the rest get 409. A seller cannot buy their own listing.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Sold, money moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not a recycler, or this is your own listing
        '404':
          description: Listing not found
        '409':
          description: Already sold or withdrawn
  /api/v1/listings/{id}/cancel:
    post:
      tags: [Marketplace]
      summary: Withdraw a listing (seller)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Withdrawn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not your listing
        '409':
          description: Already sold or already withdrawn
  /api/v1/wallet:
    get:
      tags: [Marketplace]
      summary: Wallet balance and payment history
      description: >
        Works for any role. A citizen sees credits from sales, a recycler sees debits from
        purchases. Green Points are returned alongside the money balance so one call powers a
        profile screen.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Balance plus paginated transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Missing, malformed or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    CreateListingRequest:
      type: object
      required: [price]
      properties:
        detectionId:
          type: string
          format: uuid
          description: Optional. Copies material, weight, image and description from the scan.
        material:
          type: string
          maxLength: 80
          example: Plastic
          description: Required when no detectionId is given.
        weightKg:
          type: number
          example: 15
          description: Required when no detectionId is given.
        price:
          type: number
          minimum: 1
          example: 120
        description:
          type: string
          maxLength: 400
        location:
          type: string
          maxLength: 160
          example: Shibpur, Howrah
        imageUrl:
          type: string
          maxLength: 500
    Listing:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [OPEN, SOLD, CANCELLED]
        material: { type: string, example: PET Bottle x13 }
        weightKg: { type: number, example: 15.0 }
        price: { type: number, example: 120 }
        pricePerKg:
          type: number
          nullable: true
          description: Derived, so a buyer can compare listings.
          example: 8.0
        currency: { type: string, example: INR }
        description: { type: string, nullable: true }
        imageUrl:
          type: string
          nullable: true
          description: The scan photo when available, otherwise a placeholder.
        location: { type: string, nullable: true }
        seller:
          $ref: '#/components/schemas/ListingParty'
        buyer:
          allOf:
            - $ref: '#/components/schemas/ListingParty'
          nullable: true
        mine:
          type: boolean
          description: True when the caller is the seller.
        createdAt: { type: string, format: date-time }
        soldAt: { type: string, format: date-time, nullable: true }
    ListingParty:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        role: { type: string }
    ListingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }
        hasMore: { type: boolean }
    Wallet:
      type: object
      properties:
        balance: { type: number, example: 120.00 }
        currency: { type: string, example: INR }
        totalEarned: { type: number, example: 120.00 }
        totalSpent: { type: number, example: 0 }
        greenPoints:
          type: integer
          description: Reward points, returned here so one call powers a profile screen.
          example: 65
        transactions:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              type:
                type: string
                enum: [CREDIT, DEBIT]
              amount: { type: number, example: 120.00 }
              balanceAfter: { type: number, example: 120.00 }
              currency: { type: string }
              reason:
                type: string
                enum: [LISTING_SOLD, LISTING_PURCHASED]
              note: { type: string, example: Sold PET Bottle x13 to Kolkata Recyclers }
              listingId: { type: string, format: uuid }
              createdAt: { type: string, format: date-time }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        hasMore: { type: boolean }
...
security:
  - bearerAuth: []